+++ android/src/main/java/io/sentry/react/RNSentryModule.java
@@ -31,6 +31,7 @@ import java.util.UUID;
 import java.util.logging.Level;
 import java.util.logging.Logger;
 
+import io.sentry.SentryEvent;
 import io.sentry.android.core.AnrIntegration;
 import io.sentry.android.core.AppStartState;
 import io.sentry.android.core.NdkIntegration;
@@ -43,6 +44,7 @@ import io.sentry.SentryLevel;
 import io.sentry.UncaughtExceptionHandlerIntegration;
 import io.sentry.protocol.SdkVersion;
 import io.sentry.protocol.SentryException;
+import io.sentry.protocol.SentryPackage;
 import io.sentry.protocol.User;
 
 @ReactModule(name = RNSentryModule.NAME)
@@ -150,25 +152,8 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
                     // We do nothing
                 }
 
-                // Add on the correct event.origin tag.
-                // it needs to be here so we can determine where it originated from.
-                SdkVersion sdkVersion = event.getSdk();
-                if (sdkVersion != null) {
-                    String sdkName = sdkVersion.getName();
-                    if (sdkName != null) {
-                        if (sdkName.equals("sentry.javascript.react-native")) {
-                            event.setTag("event.origin", "javascript");
-                        } else if (sdkName.equals("sentry.java.android") || sdkName.equals("sentry.native")) {
-                            event.setTag("event.origin", "android");
-
-                            if (sdkName.equals("sentry.native")) {
-                                event.setTag("event.environment", "native");
-                            } else {
-                                event.setTag("event.environment", "java");
-                            }
-                        }
-                    }
-                }
+                setEventOriginTag(event);
+                addPackages(event, options.getSdkVersion());
 
                 return event;
             });
@@ -444,6 +429,49 @@ public class RNSentryModule extends ReactContextBaseJavaModule {
         if (RNSentryModule.frameMetricsAggregator != null) {
             RNSentryModule.frameMetricsAggregator.stop();
             RNSentryModule.frameMetricsAggregator = null;
+        }
+    }
+
+    private void setEventOriginTag(SentryEvent event) {
+        SdkVersion sdk = event.getSdk();
+        if (sdk != null) {
+          switch (sdk.getName()) {
+          // If the event is from capacitor js, it gets set there and we do not handle it here.
+          case "sentry.native":
+            setEventEnvironmentTag(event, "android", "native");
+            break;
+          case "sentry.java.android":
+            setEventEnvironmentTag(event, "android", "java");
+            break;
+          default:
+            break;
+          }
+        }
+    }
+
+    private void setEventEnvironmentTag(SentryEvent event, String origin, String environment) {
+        event.setTag("event.origin", origin);
+        event.setTag("event.environment", environment);
+    }
+
+    private void addPackages(SentryEvent event, SdkVersion sdk) {
+        SdkVersion eventSdk = event.getSdk();
+        if (eventSdk != null && eventSdk.getName().equals("sentry.javascript.react-native") && sdk != null) {
+            List<SentryPackage> sentryPackages = sdk.getPackages();
+            if (sentryPackages != null) {
+                for (SentryPackage sentryPackage : sentryPackages) {
+                    eventSdk.addPackage(sentryPackage.getName(), sentryPackage.getVersion());
+                }
+            }
+
+            List<String> integrations = sdk.getIntegrations();
+            if (integrations != null) {
+                for (String integration : integrations) {
+                    eventSdk.addIntegration(integration);
+                }
+            }
+
+            event.setSdk(eventSdk);
         }
     }
 }