+++ src/js/index.ts
@@ -13,7 +13,6 @@ export {
   User,
 } from "@sentry/types";
 
-import { addGlobalEventProcessor } from "@sentry/core";
 export {
   addGlobalEventProcessor,
   addBreadcrumb,
@@ -80,32 +79,5 @@ export {
   RoutingInstrumentation,
   ReactNavigationTransactionContext,
 } from "./tracing";
-
-/**
- * Adds the sdk info. Make sure this is called after @sentry/react's so this is the top-level SDK.
- */
-function createReactNativeEventProcessor(): void {
-  if (addGlobalEventProcessor) {
-    addGlobalEventProcessor((event) => {
-      event.platform = event.platform || "javascript";
-      event.sdk = {
-        ...event.sdk,
-        name: SDK_NAME,
-        packages: [
-          ...((event.sdk && event.sdk.packages) || []),
-          {
-            name: "npm:@sentry/react-native",
-            version: SDK_VERSION,
-          },
-        ],
-        version: SDK_VERSION,
-      };
-
-      return event;
-    });
-  }
-}
-
-createReactNativeEventProcessor();
 
 export { Integrations, SDK_NAME, SDK_VERSION };