+++ src/js/integrations/index.ts
@@ -2,3 +2,5 @@ export { DebugSymbolicator } from "./debugsymbolicator";
 export { DeviceContext } from "./devicecontext";
 export { ReactNativeErrorHandlers } from "./reactnativeerrorhandlers";
 export { Release } from "./release";
+export { EventOrigin } from "./eventorigin";
+export { SdkInfo } from "./sdkinfo";